Fix hiding files for CLs with large changes
We need to wait until all files are fetched from the server before we
can start hiding them on the frontend. We can do this by waiting until
the 'Show all files' button is hidden on the frontend - which indicates
that the final payload from the server has been received.