Remove while-loop when waiting for paginated data
Prior, a while-loop was used to ping the controlRow DOM checking to see
if it was marked invisible yet. When the controlRow has the class
'invisible' added, we can ge sure that the paginated file list is
finally received by the frontend which then allows us to begin
filtering. However, this while loop is not ideal as it might hog
resources, especially if fetching data from the backend takes a long
time.

This commit introduces a mutation observer to watch for changes in the
DOM tree and returns a fulfilled promise once the controlRow element is
invisible.